# server/dist/swagger/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
...
            "ErrorResponseDto": {
                "properties": {
                    "message": {
                        "type": "string",
                        "description": "Error message",
                        "example": "Invalid request parameters"
                    },
                    "statusCode": {
                        "type": "number",
                        "format": "double",
                        "description": "HTTP status code",
                        "example": 400
                    },
                    "name": {
                        "type": "string",
                        "description": "Error name/type",
                        "example": "ValidationError"
                    }
                },
                "required": [
                    "message",
                    "statusCode",
                    "name"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT token obtained from /auth/login"
            }
        }
    },
    "info": {
        "title": "coown-api",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {
            "name": "Salawu Olanrewaju Joseph"
        }
    },
    "paths": {
...
        "/pools": {
            "get": {
                "operationId": "GetPools",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/PoolDto"
                                    },
                                    "type": "array"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": [
                                            {
                                                "id": "550e8400-e29b-41d4-a716-446655440000",
                                                "property": {
                                                    "id": "550e8400-e29b-41d4-a716-446655440001",
                                                    "title": "3 Bedroom Luxury Apartment in Lekki",
                                                    "location": "Lekki Phase 1, Lagos",
                                                    "price": 15000000,
                                                    "type": "apartment",
                                                    "status": "available"
                                                },
                                                "creator": {
                                                    "id": "550e8400-e29b-41d4-a716-446655440002",
                                                    "firstName": "John",
                                                    "lastName": "Doe",
                                                    "email": "john.doe@example.com",
                                                    "role": "user",
                                                    "verified": true,
                                                    "isActive": true
                                                },
                                                "name": "Lagos Luxury Apartment Co-Own",
                                                "target_amount": 5000000,
                                                "raised_amount": 3250000,
                                                "deadline": "2024-12-31T00:00:00.000Z",
                                                "status": "active",
                                                "is_public": false,
                                                "created_at": "2024-01-15T10:30:00.000Z",
                                                "updatedAt": "2024-01-15T10:30:00.000Z"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Invalid filter parameters",
                                            "statusCode": 400,
                                            "name": "ValidationError"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Pools"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "query",
                        "name": "creatorId",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "isPublic",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ]
            },
            "post": {
                "operationId": "CreatePool",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PoolDto"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Pool Created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PoolDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Invalid pool data or property not found",
                                            "statusCode": 400,
                                            "name": "ValidationError"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Authentication required",
                                            "statusCode": 401,
                                            "name": "UnauthorizedError"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Pools"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePoolDto"
                            }
                        }
                    }
                }
            }
        },
...
        "/pools/{id}": {
            "get": {
                "operationId": "GetPoolById",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PoolDto"
                                        }
                                    ],
                                    "nullable": true
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pool Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Pool not found",
                                            "statusCode": 404,
                                            "name": "NotFoundError"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Pools"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdatePool",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Pool"
                                        }
                                    ],
                                    "nullable": true
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Pools"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Partial_Pool_"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeletePool",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Pools"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/pools/{id}/join": {
            "post": {
                "operationId": "JoinPool",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PoolMemberDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "id": "550e8400-e29b-41d4-a716-446655440003",
                                            "pool_id": "550e8400-e29b-41d4-a716-446655440000",
                                            "user_id": "550e8400-e29b-41d4-a716-446655440002",
                                            "declared_amount": 250000,
                                            "paid_amount": 0,
                                            "ownership_pct": 25,
                                            "joined_at": "2024-01-15T10:30:00.000Z"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Successfully joined pool",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PoolMemberDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Invalid amount or already a member",
                                            "statusCode": 400,
                                            "name": "ValidationError"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Authentication required",
                                            "statusCode": 401,
                                            "name": "UnauthorizedError"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pool Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Pool not found",
                                            "statusCode": 404,
                                            "name": "NotFoundError"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Pools"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JoinPoolDto"
                            }
                        }
                    }
                }
            }
        },
        "/pools/{id}/dashboard": {
            "get": {
                "operationId": "GetPoolDashboard",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PoolDashboardDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pool Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Pool not found",
                                            "statusCode": 404,
                                            "name": "NotFoundError"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Pools"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/pools/{id}/members": {
            "get": {
                "operationId": "GetPoolMembers",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/PoolMemberDto"
                                    },
                                    "type": "array"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": [
                                            {
                                                "id": "550e8400-e29b-41d4-a716-446655440003",
                                                "pool_id": "550e8400-e29b-41d4-a716-446655440000",
                                                "user_id": "550e8400-e29b-41d4-a716-446655440002",
                                                "declared_amount": 250000,
                                                "paid_amount": 0,
                                                "ownership_pct": 25,
                                                "joined_at": "2024-01-15T10:30:00.000Z"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pool Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Pool not found",
                                            "statusCode": 404,
                                            "name": "NotFoundError"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Pools"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/pools/{id}/users": {
            "get": {
                "operationId": "GetPoolUsers",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pool Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Pool not found",
                                            "statusCode": 404,
                                            "name": "NotFoundError"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Pools"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}